hotfix: statsd: return if buffer length is 0.
This moves the buffer length check below the creation of the buffer
string and the emptying of the buffer array. I believe I was hitting a
race condition where another request completed and cleared the buffer
while this function was executing, causing the statsd buffer refence to
empty during processing.